convert react-fiber-react-18 to typescript
