--- packages/react-fiber-react-18/module.js
@@ -1,9 +1,0 @@
-/* eslint-disable global-require */
-if (['test', 'development'].includes(process.env.NODE_ENV))
-{
-    module.exports = require('./dist/index.es-dev');
-}
-else
-{
-    module.exports = require('./dist/index.es');
-}
