+++ packages/react-fiber-react-18/rollup.config.mjs
@@ -1,8 +1,6 @@
-import alias from '@rollup/plugin-alias';
-
 import {
     getBuildFormat,
-    getRollupConfig,
+    getRollupTSConfig,
     isProductionBuild,
 } from '../../shared/getRollupConfig.mjs';
 
@@ -21,33 +19,25 @@ let builds;
 if (format)
 {
     builds = [
-        getRollupConfig(`../react-fiber/v18/dist/index.${format}${buildType}.js`, format, {
-            beforePlugins: [
-                alias({
-                    entries: {
-                        '@react-spring/animated':
-                            '../../shared/react-spring-create-host.js',
-                    },
-                }),
-            ],
-            external,
-        }),
+        getRollupTSConfig(
+            `../react-fiber/v18/dist/index.${format}${buildType}.js`,
+            format,
+            {
+                external,
+            }
+        ),
     ];
 }
 else
 {
     builds = ['cjs', 'es'].map((format) =>
-        getRollupConfig(`../react-fiber/v18/dist/index.${format}${buildType}.js`, format, {
-            beforePlugins: [
-                alias({
-                    entries: {
-                        '@react-spring/animated':
-                            '../../shared/react-spring-create-host.js',
-                    },
-                }),
-            ],
-            external,
-        })
+        getRollupTSConfig(
+            `../react-fiber/v18/dist/index.${format}${buildType}.js`,
+            format,
+            {
+                external,
+            }
+        )
     );
 }
