+++ packages/react-fiber-react-18/src/data.json
@@ -1,0 +1,4 @@
+{
+  "REACT_DOM_VERSION": "18.0.0",
+  "PACKAGE_NAME": "@pixi/react-fiber-react-18"
+}
